review: Move `ProcessedToDeviceEvent` to crypto types mod
