+++ crates/matrix-sdk-crypto/src/machine/tests/megolm_sender_data.rs
@@ -17,7 +17,7 @@ use std::{fmt::Debug, iter, pin::Pin};
 use assert_matches::assert_matches;
 use futures_core::Stream;
 use futures_util::{FutureExt, StreamExt};
-use matrix_sdk_common::deserialized_responses::ProcessedToDeviceEvent;
+use crate::types::ProcessedToDeviceEvent;
 use matrix_sdk_test::async_test;
 use ruma::{room_id, user_id, RoomId, TransactionId, UserId};
 use serde::Serialize;
